// ClientApp/generated/StatusMessageModel.ts
/*  GENERTAED automatically using Models.tst. DO NOT CHANGE HERE */

//Generated from class components.revLogin.StatusMessageModel
export interface StatusMessageModel {
  version: string;
  revConnect: string;
  signOutUrl: string;
  openIdProviders: { [key: string]: string };
  restrictedNewWorkspace: boolean;
  sessionId: string;
  jwt: string;
}